event system enemy

+++ Game/TileMap.cpp
@@ -27,13 +27,13 @@ bool nc::TileMap::Create(Scene* scene)
             // i = 0, 1, 2, 3, 4, 5, 6, 7, 8
             //     0, 1, 2, 3, 4
             //     0, 1, 2, 3, 4
-            float x = i % m_numX;
+            float x = static_cast<float>(i % m_numX);
 
             // i = 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10
             //     0, 0, 0, 0, 0
             //     1, 1, 1, 1, 1
 
-            float y = i / m_numX;
+            float y = static_cast<float>(i / m_numX);
             gameObject->m_transform.position = m_start + (nc::Vector2{ x, y } * m_size);
 
             scene->AddGameObject(gameObject);
